폐기/requirements.txt
neo4j==5.20.0
google-generativeai
pdfplumber
numpy
scikit-learn
sentence-transformers==3.0.1
requests
beautifulsoup4
